requirements.txt
groq
python-dotenv
langchain
ollama
chromadb
tiktoken
langchain-groq
langchain-community
bs4
langchain_ollama
langchain_huggingface
apify-client